pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pii-guard"
version = "0.1.0"
description = "Lokaler Datenschutz-Filter für KI-Coding-Tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Thomas Körting", email = "tk@b-imtec.de" },
]
keywords = ["pii", "privacy", "claude-code", "llm", "data-protection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
    "faker>=24.0",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "pytest-cov>=5.0",
]
gitleaks = [
    "pre-commit>=3.7",
]

[project.scripts]
pii-guard = "pii_guard.cli:main"

[project.urls]
Homepage = "https://github.com/tkoerting/pii-guard"
Repository = "https://github.com/tkoerting/pii-guard"
Issues = "https://github.com/tkoerting/pii-guard/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pii_guard"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
